requirements.txt
streamlit==1.34.0
tensorflow==2.12.0
numpy==1.24.3
Pillow==9.5.0
gdown
